unify type generation
