+++ packages/astro/tsconfig.json
@@ -1,6 +1,4 @@
 {
-  "extends": "../../tsconfig.types.json",
-  "compilerOptions": { "outDir": "dist", "jsx": "react-jsx" },
-  "include": ["lib/**/*.js", "index.js"],
-  "exclude": ["node_modules/"]
+	"extends": "../../tsconfig.json",
+	"compilerOptions": { "jsx": "react-jsx" }
 }
