--- packages/editor/.gitignore
@@ -1,1 +1,0 @@
-index.d.ts
