+++ packages/lezer/index.js
@@ -1,0 +1,2 @@
+export { OrgParser, parser } from './lib/index.js'
+export { tags } from './lib/nodes.js'
