+++ packages/oast-to-prose/index.js
@@ -1,0 +1,5 @@
+/**
+ * @typedef {import('./lib/index.js').Options} Options
+ */
+
+export { schema, toProse } from './lib/index.js'
