+++ packages/oast-to-prose/package.json
@@ -1,33 +1,29 @@
 {
-  "name": "oast-to-prose",
-  "version": "1.2.0",
-  "description": "Transform OAST to ProseMirror Document",
-  "author": "Xiaoxing Hu <hi@xiaoxing.dev>",
-  "license": "MIT",
-  "main": "lib/index.js",
-  "type": "module",
-  "types": "dist/index.d.ts",
-  "homepage": "https://github.com/orgapp/orgajs/tree/main/packages/oast-to-prose#readme",
-  "repository": {
-    "type": "git",
-    "url": "https://github.com/orgapp/orgajs.git",
-    "directory": "packages/oast-to-prose"
-  },
-  "files": [
-    "lib",
-    "dist"
-  ],
-  "scripts": {
-    "clean": "tsc --build --clean",
-    "build": "tsc --build",
-    "test": "echo \"\\033[31mTODO: fix tests\""
-  },
-  "keywords": [],
-  "devDependencies": {
-    "orga": "workspace:^",
-    "vfile": "^6.0.3"
-  },
-  "dependencies": {
-    "prosemirror-model": "^1.19.3"
-  }
+	"name": "oast-to-prose",
+	"version": "1.2.0",
+	"description": "Transform OAST to ProseMirror Document",
+	"author": "Xiaoxing Hu <hi@xiaoxing.dev>",
+	"license": "MIT",
+	"type": "module",
+	"homepage": "https://github.com/orgapp/orgajs/tree/main/packages/oast-to-prose#readme",
+	"repository": {
+		"type": "git",
+		"url": "https://github.com/orgapp/orgajs.git",
+		"directory": "packages/oast-to-prose"
+	},
+	"files": [
+		"lib",
+		"index.js",
+		"index.d.ts",
+		"index.d.ts.map"
+	],
+	"scripts": {},
+	"keywords": [],
+	"devDependencies": {
+		"orga": "workspace:^",
+		"vfile": "^6.0.3"
+	},
+	"dependencies": {
+		"prosemirror-model": "^1.19.3"
+	}
 }
