+++ packages/orga-build/tsconfig.json
@@ -1,6 +1,3 @@
 {
-	"extends": "../../tsconfig.esm.json",
-	"compilerOptions": {
-		"module": "esnext"
-	}
+	"extends": "../../tsconfig.json"
 }
