--- packages/orgx/.gitignore
@@ -1,1 +1,0 @@
-*.d.ts
