+++ packages/orgx/tsconfig.json
@@ -1,6 +1,3 @@
 {
-  "extends": "../../tsconfig.types.json",
-  "compilerOptions": { "outDir": "dist", "jsx": "react-jsx" },
-  "include": ["lib/**/*", "index.js"],
-  "exclude": ["node_modules/"]
+	"extends": "../../tsconfig.json"
 }
