+++ packages/react/tsconfig.json
@@ -1,6 +1,3 @@
 {
-  "extends": "../../tsconfig.types.json",
-  "compilerOptions": { "outDir": "dist" },
-  "include": ["lib/**/*.js", "index.js"],
-  "exclude": ["node_modules/"]
+	"extends": "../../tsconfig.json"
 }
