+++ packages/reorg-prose/index.js
@@ -1,0 +1,2 @@
+export { schema } from 'oast-to-prose'
+export { default as default } from './lib/index.js'
