+++ packages/reorg-prose/package.json
@@ -1,35 +1,32 @@
 {
-  "name": "@orgajs/reorg-prose",
-  "version": "1.2.0",
-  "description": "",
-  "author": "Xiaoxing Hu <hi@xiaoxing.dev>",
-  "license": "MIT",
-  "main": "lib/index.js",
-  "type": "module",
-  "types": "dist/index.d.ts",
-  "homepage": "https://github.com/orgapp/orgajs/tree/main/packages/reorg-prose#readme",
-  "repository": {
-    "type": "git",
-    "url": "https://github.com/orgapp/orgajs.git",
-    "directory": "packages/reorg-prose"
-  },
-  "files": [
-    "lib",
-    "dist"
-  ],
-  "scripts": {
-    "clean": "tsc --build --clean",
-    "build": "tsc --build"
-  },
-  "keywords": [],
-  "dependencies": {
-    "oast-to-prose": "workspace:^"
-  },
-  "devDependencies": {
-    "@types/unist": "^3.0.3",
-    "orga": "workspace:^",
-    "prosemirror-model": "^1.19.3",
-    "unified": "^11.0.5",
-    "vfile": "^6.0.3"
-  }
+	"name": "@orgajs/reorg-prose",
+	"version": "1.2.0",
+	"description": "",
+	"author": "Xiaoxing Hu <hi@xiaoxing.dev>",
+	"license": "MIT",
+	"type": "module",
+	"homepage": "https://github.com/orgapp/orgajs/tree/main/packages/reorg-prose#readme",
+	"repository": {
+		"type": "git",
+		"url": "https://github.com/orgapp/orgajs.git",
+		"directory": "packages/reorg-prose"
+	},
+	"files": [
+		"lib",
+		"index.js",
+		"index.d.ts",
+		"index.d.ts.map"
+	],
+	"scripts": {},
+	"keywords": [],
+	"dependencies": {
+		"oast-to-prose": "workspace:^"
+	},
+	"devDependencies": {
+		"@types/unist": "^3.0.3",
+		"orga": "workspace:^",
+		"prosemirror-model": "^1.19.3",
+		"unified": "^11.0.5",
+		"vfile": "^6.0.3"
+	}
 }
